Updated makefile to correctly set platform for Proxmox Debian installs
The default installation returns "unknown" for Proxmox installations when `uname -i` is called. To address this, i have hard set the platform to x86_64 to ensure this builds on Proxmox installs.

NOTE: My bash scripting is amateur at best hence the hard property switch in this change. The next step is to replace this change with a switch to look for platform in 'dpkg --print architecture` or similar call to detect Proxmox installations.